+++ app/Http/Controllers/APIs/StoreController.php
@@ -4,47 +4,70 @@ namespace App\Http\Controllers\APIs;
 use Illuminate\Http\Request;
 use App\Models\Store;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Validator;
 
 class StoreController extends Controller
 {
-    public function index()
+    /* -------------------------------------------get all store ------------------------------------------------ */
+    public function getAll()
     {
-        return response()->json(Store::get(), 200);
+        $store = Store::get()->toJson();
+        return response($store, 200);
     }
 
-    public function create()
+    /* ------------------------------------- create an store -------------------------------------- */
+    public function create(Request $request): \Illuminate\Http\JsonResponse
     {
-        //
+
+        $data = $request->all();
+        //validator or request validator
+        $validator = Validator::make($data, [
+            'name' => 'required|max:255',
+        ]);
+
+        if ($validator->fails()) {
+            return response()->json(['message' => $validator->errors()], 'Validation Error');
+        }
+
+        $store = Store::create($data);
+        return response()->json(["message" => "store record created"], 201);
     }
 
-    public function store(Request $request)
+
+    /* -------------------------------------get one store -------------------------------------- */
+    public function getOne($id)
     {
-        $store = Store::create($request->all());
-        return response()->json($store, 201);
+        if (Store::where('id', $id)->exists()) {
+            $store = Store::where('id', $id)->get()->toJson();
+            return response($store, 200);
+        } else {
+            return response()->json(["message" => "store not found"], 404);
+        }
     }
 
-    public function show($id)
+    /* -------------------------------------update one store -------------------------------------- */
+    public function update(Request $request, $id): \Illuminate\Http\JsonResponse
     {
-        $store = Store::find($id);
-        return response()-json($store, 200);
+        if (Store::where('id', $id)->exists()) {
+            $store = Store::find($id);
+            $store->name = is_null($request->name) ? $store->name : $request->name;
+            $store->save();
+
+            return response()->json(["message" => "store updated successfully"], 200);
+        } else {
+            return response()->json(["message" => "store not found"], 404);
+        }
     }
 
-    public function edit($id)
+    /* -------------------------------------delete store -------------------------------------- */
+    public function delete($id): \Illuminate\Http\JsonResponse
     {
-        //
-    }
-
-    public function update(Request $request, $id)
-    {
-        $store = Store::find($id);
-        $store->update($request->all());
-	    return response()->json($store, 200);
-    }
-
-    public function destroy($id)
-    {
-        $store = Store::find($id);
-        $store->delete();
-        return response()->json(null, 204);
+        if(Store::where('id', $id)->exists()) {
+            $store = Store::find($id);
+            $store->delete();
+            return response()->json(["message" => "store record deleted"], 202);
+        } else {
+            return response()->json(["message" => "store not found"], 404);
+        }
     }
 }
